Extract shared IncidentArchive from CrashArchive and HangArchive
