+++ Sources/Scout/Core/Diagnostics/Crash/CrashArchive.swift
@@ -7,13 +7,14 @@
 
 import Foundation
 
-/// Handles file-based persistence and recovery of crash reports.
-///
-/// Crash data is written synchronously to disk to ensure it's saved
-/// before the application terminates.
-///
+// Crash data is written synchronously to disk to ensure it's saved before the
+// application terminates.
 struct CrashArchive {
-    let directory: URL
+    private let archive: IncidentArchive<CrashInfo>
+
+    init(directory: URL) {
+        archive = IncidentArchive(directory: directory, pathExtension: "crash", persist: logCrash)
+    }
 
     static let system = CrashArchive(
         directory: FileManager.default
@@ -23,46 +24,10 @@ struct CrashArchive {
     )
 
     func write(_ crash: CrashInfo) {
-        let encoder = JSONEncoder()
-        encoder.dateEncodingStrategy = .iso8601
-
-        guard let data = try? encoder.encode(crash) else {
-            return
-        }
-
-        try? FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-
-        let fileName = "\(UUID().uuidString).crash"
-        let fileURL = directory.appendingPathComponent(fileName)
-
-        try? data.write(to: fileURL, options: .atomic)
+        archive.write(crash)
     }
 
     func flush(deviceID: UUID) async {
-        guard let files = try? FileManager.default.contentsOfDirectory(at: directory, includingPropertiesForKeys: nil)
-        else {
-            return
-        }
-
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .iso8601
-
-        for file in files where file.pathExtension == "crash" {
-            guard let data = try? Data(contentsOf: file), let crash = try? decoder.decode(CrashInfo.self, from: data)
-            else {
-                try? FileManager.default.removeItem(at: file)
-                continue
-            }
-
-            do {
-                let id = UUID(uuidString: file.deletingPathExtension().lastPathComponent) ?? UUID()
-                try await persistentContainer.performBackgroundTask { context in
-                    try logCrash(crash, id: id, deviceID: deviceID, context: context)
-                }
-                try FileManager.default.removeItem(at: file)
-            } catch {
-                print("Failed to process crash: \(error.localizedDescription)")
-            }
-        }
+        await archive.flush(deviceID: deviceID)
     }
 }
